requirements.txt
playwright==1.40.0
playwright-stealth==1.0.6
requests==2.31.0
python-telegram-bot==20.7
aiohttp==3.9.1
fake-useragent==1.4.0
python-dotenv==1.0.0